+++ Models/Models.cs
@@ -98,7 +98,21 @@ namespace UIFramework
 					}
 					catch (Exception ex)
 					{
-						Debug.Log($"Error savinvg category {model.Identifier} for mod {Instance.Info.Name}: {ex.Message}", false, 2);
+						Debug.Log($"Error saving category {model.Identifier} for mod {Instance.Info.Name}: {ex.Message}", false, 2);
+					}
+				}
+			}
+			public override void DiscardAction()
+			{
+				foreach (IModelable model in SubModels)
+				{
+					try
+					{
+						model.DiscardAction();
+					}
+					catch (Exception ex)
+					{
+						Debug.Log($"Error loading category {model.Identifier} for mod {Instance.Info.Name}: {ex.Message}", false, 2);
 					}
 				}
 			}