+++ Assets/ThunderstorePackage/Template/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 0.5.1
 - Added non-contiguous enum support
+- Added discard button
 
 # Version 0.5.0
 - Added support for enum display names